backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0   # JWT token implementation for DRF
dj-rest-auth[with_social]==7.0.1     # REST API authentication utils
django-allauth==65.9.0
drf_spectacular==0.28.0     # API documentation
django-cors-headers==4.7.0   # Enable Cross-Origin Resource Sharing (CORS)
psycopg[binary]==3.2.6 # A PostgreSQL  Adapter
pytest-django==4.10.0 # a testing library
python-dotenv==1.1.0 # for loading environment variables from .env files
colorlog==6.9.0 # for colored terminal output
requests==2.32.4
oauthlib==3.2.2

# FAERS analysis requirements: https://github.com/bgbg/faers_analysis/blob/main/requirements.txt
tqdm==4.67.1
matplotlib==3.10.3
numpy==2.2.4
openpyxl==3.1.5 
pandas==2.2.3
seaborn==0.13.2
statsmodels==0.14.4
luigi==3.6.0